fix 保证金

+++ shop-dao/src/main/java/com/masiis/shop/dao/platform/product/PfSkuAgentMapper.java
@@ -11,17 +11,16 @@ import com.masiis.shop.dao.po.PfSkuAgent;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Map;
 
 @Repository
 public interface PfSkuAgentMapper {
     int deleteByPrimaryKey(Integer id);
 
     int insert(PfSkuAgent record);
-    Double selectMaxBail(Integer skuId);
-    Double selectMinBail(Integer skuId);
+    BigDecimal selectMaxBail(Integer skuId);
+    BigDecimal selectMinBail(Integer skuId);
 
     PfSkuAgent selectByPrimaryKey(Integer id);
